id由long改为int 完善Redis登录校验 完成更改密码功能 完成查询菜品存入Redis缓存功能

+++ src/main/java/com/holy/domain/po/Dish.java
@@ -13,10 +13,10 @@ import java.io.Serializable;
 @EqualsAndHashCode(callSuper = false)
 @Accessors(chain = true)
 @TableName("dish")
-public class Dish implements Serializable {
+public class Dish implements Serializable{
 
     @TableId(value = "id", type = IdType.AUTO)
-    private long id;
+    private int id;
 
     // 菜品名称
     private String dishName;
